support indexed-v1 feed format
