+++ package.json
@@ -23,11 +23,11 @@
     "pull-cat": "^1.1.11",
     "pull-notify": "^0.1.1",
     "pull-stream": "^3.6.14",
-    "ssb-bfe": "^3.3.0",
-    "ssb-db2": ">=3.0.0 <=5",
-    "ssb-keys": "^8.4.0",
+    "ssb-bfe": "^3.6.0",
+    "ssb-db2": ">=3.0.0 <=6",
+    "ssb-keys": "^8.5.0",
     "ssb-ref": "^2.16.0",
-    "ssb-uri2": "^2.0.2"
+    "ssb-uri2": "^2.1.0"
   },
   "devDependencies": {
     "c8": "^7.11.0",
@@ -37,7 +37,7 @@
     "rimraf": "^3.0.2",
     "secret-stack": "^6.4.0",
     "ssb-bendy-butt": "^1.0.1",
-    "ssb-db2": "^5.1.0",
+    "ssb-db2": "^6.1.1",
     "ssb-caps": "^1.1.0",
     "tap-arc": "^0.3.5",
     "tape": "^5.5.3"
