+++ test/api.js
@@ -131,7 +131,7 @@ tape('findOrCreate() a subfeed under a sub meta feed', (t) => {
           (f) => f.feedpurpose === 'index',
           {
             feedpurpose: 'index',
-            feedformat: 'classic',
+            feedformat: 'indexed-v1',
             metadata: { query: 'foo' },
           },
           (err, f) => {
@@ -139,7 +139,10 @@ tape('findOrCreate() a subfeed under a sub meta feed', (t) => {
             t.error(err, 'no err')
             t.equals(f.feedpurpose, 'index', 'it is the index subfeed')
             t.equals(f.metadata.query, 'foo', 'query is okay')
-            t.true(f.subfeed.endsWith('.ed25519'), 'is a classic feed')
+            t.true(
+              f.subfeed.startsWith('ssb:feed/indexed-v1/'),
+              'feed format is indexed-v1'
+            )
 
             t.end()
           }
@@ -164,7 +167,7 @@ test('findById and findByIdSync', (t) => {
       ])
       t.equals(details.feedpurpose, 'index')
       t.equals(details.metafeed, testIndexesMF.keys.id)
-      t.equals(details.feedformat, 'classic')
+      t.equals(details.feedformat, 'indexed-v1')
 
       t.throws(
         () => {
@@ -228,7 +231,7 @@ test('restart sbot', (t) => {
       const details = sbot.metafeeds.findByIdSync(testIndexFeed)
       t.equals(details.feedpurpose, 'index')
       t.equals(details.metafeed, testIndexesMF.keys.id)
-      t.equals(details.feedformat, 'classic')
+      t.equals(details.feedformat, 'indexed-v1')
 
       sbot.metafeeds.getRoot((err, mf) => {
         t.error(err, 'no err')
